Update BatchReport tests to use rtl
